+++ src/frontend/custom_widgets/tracker_management.py
@@ -711,9 +711,6 @@ class HunoTrackerEdit(TrackerEditBase):
         internal_lbl = QLabel("Internal", self)
         self.internal = QCheckBox(self)
 
-        stream_optimized_lbl = QLabel("Stream Optimized", self)
-        self.stream_optimized = QCheckBox(self)
-
         image_width_lbl = QLabel("Image Width", self)
         self.image_width = QSpinBox(self)
         self.image_width.setRange(300, 2000)
@@ -722,7 +719,6 @@ class HunoTrackerEdit(TrackerEditBase):
         self.add_pair_to_layout(api_key_lbl, self.api_key)
         self.add_pair_to_layout(anonymous_lbl, self.anonymous)
         self.add_pair_to_layout(internal_lbl, self.internal)
-        self.add_pair_to_layout(stream_optimized_lbl, self.stream_optimized)
         self.add_pair_to_layout(image_width_lbl, self.image_width)
         self.add_screen_shot_settings()
 
@@ -737,7 +733,6 @@ class HunoTrackerEdit(TrackerEditBase):
         self.api_key.setText(tracker_data.api_key if tracker_data.api_key else "")
         self.anonymous.setChecked(bool(tracker_data.anonymous))
         self.internal.setChecked(bool(tracker_data.internal))
-        self.stream_optimized.setChecked(bool(tracker_data.stream_optimized))
         self.image_width.setValue(tracker_data.image_width)
         if self.screen_shot_settings:
             self.screen_shot_settings.load_settings(
@@ -759,9 +754,6 @@ class HunoTrackerEdit(TrackerEditBase):
         self.config.settings.trackers.huno.api_key = self.api_key.text().strip()
         self.config.settings.trackers.huno.anonymous = self.anonymous.isChecked()
         self.config.settings.trackers.huno.internal = self.internal.isChecked()
-        self.config.settings.trackers.huno.stream_optimized = (
-            self.stream_optimized.isChecked()
-        )
         self.config.settings.trackers.huno.image_width = self.image_width.value()
         if self.screen_shot_settings:
             col_s, col_space, row_space = self.screen_shot_settings.current_settings()
@@ -807,8 +799,11 @@ class LSTTrackerEdit(TrackerEditBase):
         featured_lbl = QLabel("Featured", self)
         self.featured = QCheckBox(self)
 
-        free_lbl = QLabel("Free", self)
-        self.free = QCheckBox(self)
+        free_lbl = QLabel("Freeleech", self)
+        self.free = QSpinBox(self)
+        self.free.setRange(0, 100)
+        self.free.setSuffix("%")
+        self._disable_scrollwheel_spinbox(self.free)
 
         double_up_lbl = QLabel("Double Up", self)
         self.double_up = QCheckBox(self)
@@ -849,7 +844,7 @@ class LSTTrackerEdit(TrackerEditBase):
         self.draft_queue_opt_in.setChecked(bool(tracker_data.draft_queue_opt_in))
         self.image_width.setValue(tracker_data.image_width)
         self.featured.setChecked(bool(tracker_data.featured))
-        self.free.setChecked(bool(tracker_data.free))
+        self.free.setValue(tracker_data.free)
         self.double_up.setChecked(bool(tracker_data.double_up))
         self.sticky.setChecked(bool(tracker_data.sticky))
         if self.screen_shot_settings:
@@ -883,7 +878,7 @@ class LSTTrackerEdit(TrackerEditBase):
         )
         self.config.settings.trackers.lst.image_width = self.image_width.value()
         self.config.settings.trackers.lst.featured = self.featured.isChecked()
-        self.config.settings.trackers.lst.free = self.free.isChecked()
+        self.config.settings.trackers.lst.free = self.free.value()
         self.config.settings.trackers.lst.double_up = self.double_up.isChecked()
         self.config.settings.trackers.lst.sticky = self.sticky.isChecked()
         if self.screen_shot_settings: